fix(state_log_viewer): 최신 버전의 libfmt 사용 시 오류 발생하던 문제 수정

+++ src/state_log_viewer.cpp
@@ -132,7 +132,7 @@ public:
             int i = 0;
             for (auto &query: transaction->queries()) {
                 _logger->info("        [#{}] {}", i++, query->statement());
-                _logger->info("            - Type: {}", query->type());
+                _logger->info("            - Type: {}", (int) query->type());
                 _logger->info("            - Database: {}", query->database());
                 _logger->info("            - Timestamp: {}", query->timestamp());
                 _logger->info("            - AffectedRows: {}", query->affectedRows());
